reformat delete function

+++ movies.js
@@ -87,7 +87,11 @@ init();
 
 
             $('#container').on('click','#deleteMovie',function (e) {
+                var response = confirm('Are you sure you want to delete this movie?');
+                if(response == true){
                 deleteMovie(e.target.dataset.id)
+                }
+
             })
 
         //Function that allows user to edit movie rating
